fix build breaks on cf.sf.net

+++ src/priv/gfx/rndr/common/renderWindowX11.cpp
@@ -177,7 +177,7 @@ bool GN::gfx::RenderWindowX11::initInternalRenderWindow(
         }
         else if( monitor )
         {
-            scr = (int)monitor;
+            scr = (int)(intptr_t)monitor;
         } 
         else
         {
